fix(mql): The MQL serializer was not serializing curried aggregates
Add functionality that properly serializes curried aggregates into MQL.
